DjangoBackend/requirements.txt
Django~=5.0
djangorestframework~=3.14.0
djoser~=2.2.2